Conflicts:
	deps/deps.yaml
	lib/transrate/read_metrics.rb

+++ lib/transrate/read_metrics.rb
@@ -57,24 +57,30 @@ module Transrate
       # classify bam file into valid and invalid alignments
       sorted_bam = "#{File.basename(bamfile, '.bam')}.merged.sorted.bam"
       readsorted_bam = "#{File.basename(bamfile, '.bam')}.valid.sorted.bam"
-      unless File.exist? readsorted_bam
-        valid_bam, invalid_bam = split_bam bamfile
-        readsorted_bam = Samtools.readsort_bam(valid_bam, threads)
-        File.delete valid_bam
+      merged_bam = "#{File.basename(bamfile, '.bam')}.merged.bam"
+      unless File.exist? sorted_bam
+        unless File.exist? readsorted_bam
+          valid_bam, invalid_bam = split_bam bamfile
+          readsorted_bam = Samtools.readsort_bam valid_bam
+          File.delete valid_bam
+        end
       end
 
       # pass valid alignments to eXpress for assignment
       # always have to run the eXpress command to load the results
       assigned_bam = assign_and_quantify readsorted_bam
+      File.delete readsorted_bam if File.exist? readsorted_bam
 
       # merge the assigned alignments back with the invalid ones
       unless File.exist? sorted_bam
-        File.delete readsorted_bam
-        merged_bam = "#{File.basename(bamfile, '.bam')}.merged.bam"
-        Samtools.merge_bam(invalid_bam, assigned_bam, merged_bam, threads=threads)
-        File.delete invalid_bam
-        File.delete assigned_bam
-        sorted_bam = Samtools.sort_bam(merged_bam, threads)
+        unless File.exist? merged_bam
+          Samtools.merge_bam(invalid_bam, assigned_bam,
+                             merged_bam, threads=threads)
+
+          File.delete invalid_bam
+          File.delete assigned_bam
+        end
+        sorted_bam = Samtools.sort_bam(merged_bam, [4, threads].min)
         File.delete merged_bam
       end
 
@@ -154,11 +160,15 @@ module Transrate
     end
 
     def analyse_expression express_output
-      express_output.each_pair do |name, eff_count|
-        @contigs_uncovered += 1 if eff_count < 1
-        @contigs_lowcovered += 1 if eff_count < 10
+      express_output.each_pair do |name, expr|
         contig = @assembly[name]
-        contig.coverage = eff_count
+        coverage = expr[:eff_count] * @read_length / expr[:eff_len]
+        @contigs_uncovered += 1 if coverage < 1
+        @contigs_lowcovered += 1 if coverage < 10
+        contig.coverage = coverage
+        contig.eff_len = expr[:eff_len]
+        contig.eff_count = expr[:eff_count]
+        contig.tpm = expr[:tpm]
       end
     end
